Address reviewer feedback
Fixes minor issues, moves FomType and BetterDirection to new module
