+++ lib/ramble/ramble/cmd/results.py
@@ -141,7 +141,7 @@ def import_results_file(filename):
     logger.debug(filename)
 
     with open(filename) as imported_file:
-        logger.msg(f"Importing {filename}")
+        logger.msg(f"Importing results file: {filename}")
 
         ext = os.path.splitext(filename)[1]
         if ext.lower() == ".json":
